+++ src/player_internals.h
@@ -255,7 +255,7 @@ struct player_s {
 
   struct x11_s *x11;          /* for X11 video out            */
   const char   *x11_display;  /* for non-default display      */
-  unsigned long winid;        /* embedded Window ID for X11   */
+  uint32_t      winid;        /* embedded Window ID for X11   */
   int   x, y;                 /* video position               */
   int   w, h;                 /* video size                   */
   float aspect;               /* video aspect                 */
